+++ lib/amico/configuration.rb
@@ -4,6 +4,8 @@ module Amico
     attr_accessor :namespace
     attr_accessor :following_key
     attr_accessor :followers_key
+    attr_accessor :blocked_key
+    attr_accessor :reciprocated_key
     attr_accessor :page_size
 
     def configure
@@ -20,6 +22,14 @@ module Amico
 
     def followers_key
       @followers_key ||= 'followers'
+    end
+
+    def blocked_key
+      @blocked_key ||= 'blocked'
+    end
+
+    def reciprocated_key
+      @reciprocated_key ||= 'reciprocated'
     end
 
     def page_size