+++ spec/amico/relationships_spec.rb
@@ -15,10 +15,18 @@ describe Amico::Relationships do
       Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:1").should be(0)
       Amico.redis.zcard("#{Amico.namespace}:#{Amico.followers_key}:1").should be(0)
     end
+
+    it 'should add each individual to the reciprocated set if you both follow each other' do
+      Amico.follow(1, 11)
+      Amico.follow(11, 1)
+
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:1").should be(1)
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:11").should be(1)
+    end
   end
 
   describe '#unfollow' do
-    it 'should allow you to follow' do
+    it 'should allow you to unfollow' do
       Amico.follow(1, 11)
 
       Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:1").should be(1)
@@ -28,20 +36,53 @@ describe Amico::Relationships do
 
       Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:1").should be(0)
       Amico.redis.zcard("#{Amico.namespace}:#{Amico.followers_key}:11").should be(0)
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:1").should be(0)
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:11").should be(0)
     end
   end
 
-  describe '#following_count' do
-    it 'should return the correct count' do
-      Amico.follow(1, 11)
-      Amico.following_count(1).should be(1)
+  describe '#block' do
+    it 'should allow you to block someone following you' do
+      Amico.follow(11, 1)
+      Amico.block(1, 11)
+
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:11").should be(0) 
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.blocked_key}:1").should be(1) 
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:1").should be(0)
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.reciprocated_key}:11").should be(0)
+    end
+
+    it 'should allow you to block someone who is not following you' do
+      Amico.block(1, 11)
+
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:11").should be(0) 
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.blocked_key}:1").should be(1) 
+    end
+
+    it 'should not allow someone you have blocked to follow you' do
+      Amico.block(1, 11)
+
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:11").should be(0) 
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.blocked_key}:1").should be(1) 
+
+      Amico.follow(11, 1)
+
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.following_key}:11").should be(0) 
+      Amico.redis.zcard("#{Amico.namespace}:#{Amico.blocked_key}:1").should be(1) 
+    end
+
+    it 'should not allow you to block yourself' do
+      Amico.block(1, 1)
+      Amico.blocked?(1, 1).should be_false
     end
   end
 
-  describe '#followers_count' do
-    it 'should return the correct count' do
-      Amico.follow(1, 11)
-      Amico.followers_count(11).should be(1)
+  describe '#unblock' do
+    it 'should allow you to unblock someone you have blocked' do
+      Amico.block(1, 11)
+      Amico.blocked?(1, 11).should be_true
+      Amico.unblock(1, 11)
+      Amico.blocked?(1, 11).should be_false      
     end
   end
 
@@ -64,6 +105,27 @@ describe Amico::Relationships do
 
       Amico.follow(11, 1)
       Amico.follower?(1,11).should be_true
+    end
+  end
+
+  describe '#blocked?' do
+    it 'should return that someone is being blocked' do
+      Amico.block(1, 11)
+      Amico.blocked?(1, 11).should be_true
+      Amico.following?(11, 1).should be_false
+    end
+  end
+
+  describe '#reciprocated?' do
+    it 'should return true if both individuals are following each other' do
+      Amico.follow(1, 11)
+      Amico.follow(11, 1)
+      Amico.reciprocated?(1, 11).should be_true
+    end
+
+    it 'should return false if both individuals are not following each other' do
+      Amico.follow(1, 11)
+      Amico.reciprocated?(1, 11).should be_false
     end
   end
 
@@ -101,6 +163,72 @@ describe Amico::Relationships do
     end
   end
 
+  describe '#blocked' do
+    it 'should return the correct list' do
+      Amico.block(1, 11)
+      Amico.block(1, 12)
+      Amico.blocked(1).should eql(["12", "11"])
+      Amico.blocked(1, :page => 5).should eql(["12", "11"])
+    end
+
+    it 'should page correctly' do
+      add_reciprocal_followers(26, true)
+
+      Amico.blocked(1, :page => 1, :page_size => 5).size.should be(5)
+      Amico.blocked(1, :page => 1, :page_size => 10).size.should be(10)
+      Amico.blocked(1, :page => 1, :page_size => 26).size.should be(25)
+    end
+  end
+
+  describe '#reciprocated' do
+    it 'should return the correct list' do
+      Amico.follow(1, 11)
+      Amico.follow(11, 1)
+      Amico.reciprocated(1).should eql(["11"])
+      Amico.reciprocated(11).should eql(["1"])
+    end
+
+    it 'should page correctly' do
+      add_reciprocal_followers
+
+      Amico.reciprocated(1, :page => 1, :page_size => 5).size.should be(5)
+      Amico.reciprocated(1, :page => 1, :page_size => 10).size.should be(10)
+      Amico.reciprocated(1, :page => 1, :page_size => 26).size.should be(25)
+    end
+  end
+
+  describe '#following_count' do
+    it 'should return the correct count' do
+      Amico.follow(1, 11)
+      Amico.following_count(1).should be(1)
+    end
+  end
+
+  describe '#followers_count' do
+    it 'should return the correct count' do
+      Amico.follow(1, 11)
+      Amico.followers_count(11).should be(1)
+    end
+  end
+
+  describe '#blocked_count' do
+    it 'should return the correct count' do
+      Amico.block(1, 11)
+      Amico.blocked_count(1).should be(1)
+    end
+  end
+
+  describe '#reciprocated_count' do
+    it 'should return the correct count' do
+      Amico.follow(1, 11)
+      Amico.follow(11, 1)
+      Amico.follow(1, 12)
+      Amico.follow(12, 1)
+      Amico.follow(1, 13)
+      Amico.reciprocated_count(1).should be(2)
+    end
+  end
+
   describe '#following_page_count' do
     it 'should return the correct count' do
       add_reciprocal_followers
@@ -121,14 +249,38 @@ describe Amico::Relationships do
     end
   end
 
+  describe '#blocked_page_count' do
+    it 'should return the correct count' do
+      add_reciprocal_followers(26, true)
+
+      Amico.blocked_page_count(1).should be(1)
+      Amico.blocked_page_count(1, 10).should be(3)
+      Amico.blocked_page_count(1, 5).should be(5)
+    end
+  end
+
+  describe '#reciprocated_page_count' do
+    it 'should return the correct count' do
+      add_reciprocal_followers
+
+      Amico.reciprocated_page_count(1).should be(1)
+      Amico.reciprocated_page_count(1, 10).should be(3)
+      Amico.reciprocated_page_count(1, 5).should be(5)
+    end
+  end
+
   private
 
-  def add_reciprocal_followers(count = 26)
+  def add_reciprocal_followers(count = 26, block_relationship = false)
     1.upto(count) do |outer_index|
       1.upto(count) do |inner_index|
         if outer_index != inner_index
           Amico.follow(outer_index, inner_index + 1000)
           Amico.follow(inner_index + 1000, outer_index)
+          if block_relationship
+            Amico.block(outer_index, inner_index + 1000)
+            Amico.block(inner_index + 1000, outer_index)
+          end
         end
       end
     end