+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-# 1.0.0
+# 1.1.0
+
+* Added blocking to relationships
+* Added reciprocated to relationships
+
+# 1.0.0 (2012-01-11)
 
 * Initial release